requirements.txt
streamlit
pandas
requests
python-dotenv
xlsxwriter
altair
openpyxl
feedparser
